move Method
